use double for higher precision

+++ submodules/diff-gaussian-rasterization/cuda_rasterizer/backward.cu
@@ -435,8 +435,8 @@ __device__ void computeView2Gaussian_backward(
 	glm::vec3 t2 = -R_transpose * t;
 
 
-	float3 S_inv_square = {1.0f / (scale.x * scale.x + 1e-7), 1.0f / (scale.y * scale.y+ 1e-7), 1.0f / (scale.z * scale.z+ 1e-7)};
-	float C = t2.x * t2.x * S_inv_square.x + t2.y * t2.y * S_inv_square.y + t2.z * t2.z * S_inv_square.z;
+	double3 S_inv_square = {1.0f / ((double)scale.x * scale.x + 1e-7), 1.0f / ((double)scale.y * scale.y+ 1e-7), 1.0f / ((double)scale.z * scale.z+ 1e-7)};
+	double C = t2.x * t2.x * S_inv_square.x + t2.y * t2.y * S_inv_square.y + t2.z * t2.z * S_inv_square.z;
 	glm::mat3 S_inv_square_R = glm::mat3(
 		S_inv_square.x * R_transpose[0][0], S_inv_square.y * R_transpose[0][1], S_inv_square.z * R_transpose[0][2],
 		S_inv_square.x * R_transpose[1][0], S_inv_square.y * R_transpose[1][1], S_inv_square.z * R_transpose[1][2],
@@ -780,8 +780,8 @@ renderCUDA(
 									view2gaussian_j[2] * ray.x + view2gaussian_j[4] * ray.y + view2gaussian_j[5]};
 
 			// use AA, BB, CC so that the name is unique
-			float AA = ray.x * normal[0] + ray.y * normal[1] + normal[2];
-			float BB = 2 * (view2gaussian_j[6] * ray_point.x + view2gaussian_j[7] * ray_point.y + view2gaussian_j[8]);
+			double AA = ray.x * normal[0] + ray.y * normal[1] + normal[2];
+			double BB = 2 * (view2gaussian_j[6] * ray_point.x + view2gaussian_j[7] * ray_point.y + view2gaussian_j[8]);
 			float CC = view2gaussian_j[9];
 			
 			// t is the depth of the gaussian
@@ -790,7 +790,7 @@ renderCUDA(
 			if (t <= NEAR_PLANE)
 				continue;
 
-			float min_value = -(BB/AA) * (BB/4.) + CC;
+			double min_value = -(BB/AA) * (BB/4.) + CC;
 
 			float power = -0.5f * min_value;
 			if (power > 0.0f){
@@ -923,9 +923,9 @@ renderCUDA(
 			const float dL_dmin_value = dL_dpower * -0.5f;
 			// float min_value = -(BB*BB)/(4*AA) + CC;
 			// const float dL_dA = dL_dmin_value * (BB*BB)/4 *  1. / (AA*AA);
-			float dL_dA = dL_dmin_value * (BB / AA) * (BB / AA) / 4.f;
-			float dL_dB = dL_dmin_value * -BB / (2 *AA);
-			float dL_dC = dL_dmin_value * 1.0f;
+			double dL_dA = dL_dmin_value * (BB / AA) * (BB / AA) / 4.f;
+			double dL_dB = dL_dmin_value * -BB / (2 *AA);
+			double dL_dC = dL_dmin_value * 1.0f;
 
 			dL_dA += dL_dt * BB / (2 * AA * AA);
 			dL_dB += dL_dt * -1.f / (2 * AA);
